requirements.txt
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-docx>=0.8.11
